requirements.txt
Django==1.5.1
South==0.8.1
djangorestframework==2.3.5
Markdown==2.3.1
django-filter==0.6
